added tekipeps changes
