+++ client/src/layouts/HomeLayout/HomeLayout.tsx
@@ -1,0 +1,11 @@
+import { FC } from 'react';
+import HomeNavigation from './HomeNavigation';
+
+const Layout: FC = ({ children }) => (
+    <div>
+        <HomeNavigation />
+        {children}
+    </div>
+);
+
+export default Layout;
