+++ client/src/layouts/HomeLayout/HomeNavigation.module.scss
@@ -47,7 +47,7 @@
     }
 }
 
-.postRoute {
+.navLink {
     $border-width: 2px;
     border-bottom: $border-width solid $purple-theme;
     transition: border;
